requirements.txt
python-dotenv
openai
langchain
tiktoken
langchain-openai
langchain-community
gradio
chromadb
